refactor: decompose part processing
